requirements.txt
djangorestframework
django
mysqlclient
celery
python-decouple
amqp
django-celery-results
requests